Adds background/font color support for MULTIPLAYER
The only lines that changed were 49 and 50. To be exact only 49 was changed while 50 is completely new. Not sure why Github fails with "Preview changes" in this one.

This change is **untested** because I have no idea about coding ZKL. What it should do is make the MULTIPLAYER tab use background and font colors specified in SETTINGS.