Changes !voteresign threshold to 3/5+1
Used to be 1/2+1 a month ago, but people always resigned for no reason. Is currently 2/3+1, but people cry because afkers have to vote. Because afkers having to vote really is a bit of an issue I suggest 3/5+1.

Ultimately leavers and afkers should not be needed to vote, but from what I understand that requires either a redo or a new version of "spring.StartContext" to return an always-up-to-date BattleContext to prevent abuse of the system. See this commit for a nice summary of the issue: https://github.com/ZeroK-RTS/Zero-K-Infrastructure/commit/f839455503473dd83a25db4cae85d1aae384c2c4