"Report this player" -> "Contact admins" when reporting self
This commit is UNTESTED. I just randomly copied code and expect it to work.

With this change users should see a "Contact administrators" button instead of the current "Report this player" button on their own profile page. That allows players to contact admins without "reporting" themselves. Note that this commit does not change the actual text that admins see in #zkadmin.

Reason for this commit:
#zkdev
[19:30]  Rafal[ZK] i think the "Report this player" button should change to "Contact moderators" when viewing own user page
[19:30]  Rafal[ZK] see #zkadmin why